feat: Update form to generate query parameters
This commit updates the form to generate query parameters based on the input values. The form now includes an input field for the "title" attribute, and when the form is submitted, the URL is generated with the corresponding query parameter. This change improves the functionality of the form and allows for easier data retrieval on the server side.